+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to the "vcl" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.9.1]
+- Adding setting for custom varnishls path
+
+## [0.9.0]
+- Upgrading to varnishls 0.0.14
+- Updated dependencies
+
+## [0.8.2]
+- Upgrading Packages
+
+## [0.8.1]
+- New release of varnishls (0.0.13)
+
 ## [0.8.0]
 - New release of varnishls (0.0.11)